+++ just_audio/darwin/Classes/AudioPlayer.m
@@ -302,8 +302,7 @@
             @"processingState": @(_processingState),
             @"updatePosition": @((long long)1000 * _updatePosition),
             @"updateTime": @(_updateTime),
-            // TODO: buffer position
-            @"bufferedPosition": @((long long)1000 * _updatePosition),
+            @"bufferedPosition": @((long long)1000 * [self getBufferedPosition]),
             // TODO: Icy Metadata
             @"icyMetadata": (id)[NSNull null],
             @"duration": @((long long)1000 * [self getDuration]),
@@ -470,21 +469,23 @@
     }
 
     // Add next loop item if we're looping
-    if (_loopMode == loopAll) {
-        int si = [_order[0] intValue];
-        //NSLog(@"### add loop item:%d", si);
-        if (!_indexedAudioSources[si].playerItem2) {
-            [_indexedAudioSources[si] preparePlayerItem2];
-            [self addItemObservers:_indexedAudioSources[si].playerItem2];
+    if (_order.count > 0) {
+        if (_loopMode == loopAll) {
+            int si = [_order[0] intValue];
+            //NSLog(@"### add loop item:%d", si);
+            if (!_indexedAudioSources[si].playerItem2) {
+                [_indexedAudioSources[si] preparePlayerItem2];
+                [self addItemObservers:_indexedAudioSources[si].playerItem2];
+            }
+            [_player insertItem:_indexedAudioSources[si].playerItem2 afterItem:nil];
+        } else if (_loopMode == loopOne) {
+            //NSLog(@"### add loop item:%d", _index);
+            if (!_indexedAudioSources[_index].playerItem2) {
+                [_indexedAudioSources[_index] preparePlayerItem2];
+                [self addItemObservers:_indexedAudioSources[_index].playerItem2];
+            }
+            [_player insertItem:_indexedAudioSources[_index].playerItem2 afterItem:nil];
         }
-        [_player insertItem:_indexedAudioSources[si].playerItem2 afterItem:nil];
-    } else if (_loopMode == loopOne) {
-        //NSLog(@"### add loop item:%d", _index);
-        if (!_indexedAudioSources[_index].playerItem2) {
-            [_indexedAudioSources[_index] preparePlayerItem2];
-            [self addItemObservers:_indexedAudioSources[_index].playerItem2];
-        }
-        [_player insertItem:_indexedAudioSources[_index].playerItem2 afterItem:nil];
     }
 
     /* NSLog(@"after reorder: _player.items.count: ", _player.items.count); */