+++ just_audio/test/just_audio_test.dart
@@ -79,7 +79,47 @@ void runTests() {
     //expect(player.loopMode, equals(LoopMode.off));
     //expect(player.shuffleModeEnabled, equals(false));
     expect(player.automaticallyWaitsToMinimizeStalling, equals(true));
-    player.dispose();
+    await player.dispose();
+  });
+
+  test('idle-state', () async {
+    final player = AudioPlayer();
+    player.play();
+    expectState(player: player, playing: true);
+    await player.pause();
+    expectState(player: player, playing: false);
+    await player.setVolume(0.5);
+    expect(player.volume, equals(0.5));
+    await player.setSpeed(0.7);
+    expect(player.speed, equals(0.7));
+    await player.setLoopMode(LoopMode.one);
+    expect(player.loopMode, equals(LoopMode.one));
+    await player.setShuffleModeEnabled(true);
+    expect(player.shuffleModeEnabled, equals(true));
+    await player.seek(Duration(seconds: 1));
+    expectState(player: player, playing: false, position: Duration(seconds: 1));
+    final playlist = ConcatenatingAudioSource(children: []);
+    await player.setAudioSource(playlist, preload: false);
+    await playlist.addAll([
+      AudioSource.uri(
+        Uri.parse("https://foo.foo/foo.mp3"),
+        tag: 'a',
+      ),
+      AudioSource.uri(
+        Uri.parse("https://bar.bar/bar.mp3"),
+        tag: 'b',
+      ),
+      AudioSource.uri(
+        Uri.parse("https://baz.baz/baz.mp3"),
+        tag: 'c',
+      ),
+    ]);
+    await playlist.move(2, 1);
+    await playlist.removeAt(2);
+    await player.load();
+    expect(playlist.sequence.map((s) => s.tag as String).toList(),
+        equals(['a', 'c']));
+    await player.dispose();
   });
 
   test('load', () async {
@@ -94,7 +134,7 @@ void runTests() {
     expect(player.hasPrevious, equals(false));
     expect(player.sequence.length, equals(1));
     expect(player.playing, equals(false));
-    player.dispose();
+    await player.dispose();
   });
 
   test('load error', () async {
@@ -121,7 +161,7 @@ void runTests() {
       exception = e;
     }
     expect(exception != null, equals(true));
-    player.dispose();
+    await player.dispose();
   });
 
   test('control', () async {
@@ -188,7 +228,7 @@ void runTests() {
       processingState: ProcessingState.completed,
       playing: true,
     );
-    player.dispose();
+    await player.dispose();
   });
 
   test('speed', () async {
@@ -208,7 +248,7 @@ void runTests() {
     await player.setSpeed(speed2);
     await Future.delayed(period2);
     expectState(player: player, position: position2);
-    player.dispose();
+    await player.dispose();
   });
 
   test('positionStream', () async {
@@ -253,7 +293,7 @@ void runTests() {
       }
     });
     await completer.future;
-    player.dispose();
+    await player.dispose();
   });
 
   test('icyMetadata', () async {
@@ -264,7 +304,7 @@ void runTests() {
     expect(player.icyMetadata.headers.genre, equals(icyMetadata.headers.genre));
     expect((await player.icyMetadataStream.first).headers.genre,
         equals(icyMetadata.headers.genre));
-    player.dispose();
+    await player.dispose();
   });
 
   test('proxy', () async {
@@ -286,6 +326,7 @@ void runTests() {
     expect(response.headers.value(HttpHeaders.contentTypeHeader),
         equals('audio/mock'));
     await server.stop();
+    await player.dispose();
   });
 
   test('proxy0.9', () async {
@@ -310,6 +351,7 @@ void runTests() {
     await socket.close();
     expect(responseText, equals('Hello'));
     await server.stop();
+    await player.dispose();
   });
 
   test('sequence', () async {
@@ -373,6 +415,7 @@ void runTests() {
     expect(source2.sequence.map((s) => s.tag as String), equals(['e']));
     await source2.clear();
     expect(source2.sequence.map((s) => s.tag as String), equals([]));
+    await player.dispose();
   });
 
   test('detect', () async {
@@ -469,12 +512,14 @@ void runTests() {
     await player1.shuffle();
     checkIndices(player1.shuffleIndices, 5);
     expect(player1.shuffleIndices.first, equals(3));
+    await player1.dispose();
 
     final source2 = createSource();
     final player2 = AudioPlayer();
     await player2.setAudioSource(source2, initialIndex: 3);
     checkIndices(player2.shuffleIndices, 5);
     expect(player2.shuffleIndices.first, equals(3));
+    await player2.dispose();
   });
 
   test('stop', () async {
@@ -516,6 +561,7 @@ void runTests() {
     expect(player.speed, 0.7);
     expect(player.shuffleModeEnabled, true);
     expect(player.loopMode, LoopMode.one);
+    await player.dispose();
   });
 
   test('play-load', () async {
@@ -715,6 +761,7 @@ void runTests() {
     await player.setUrl('https://bar.bar/foo.mp3');
     player.stop();
     await player.setUrl('https://bar.bar/foo.mp3');
+    await player.dispose();
   });
 }