+++ just_audio_web/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.3
+
+* Fix bug when modifying playlists (insert/move).
+
+## 0.2.2
+
+* Fix bug with empty playlist.
+* Fix bug when modifying playlists.
+
 ## 0.2.1
 
 * Fix bug with play before load.